llm/requirements.txt
# TextGrad-QA Optimization
# Python 3.9+

# Core dependencies
hydra-core>=1.3.0
omegaconf>=2.3.0
pyyaml>=6.0

# LLM
llama-cpp-python>=0.2.0

# Data & NLP
datasets>=2.14.0
transformers>=4.30.0

# Utilities
tqdm>=4.65.0
numpy>=1.24.0

# Development (optional)
jupyter>=1.0.0
ipython>=8.0.0
